Bugfix: Wrong source file was fed to C2HS.
C2HS should be the step after lambda annotation expanding. But it was
fed with the input of former step, instead of the output. This was not
noted until building cairo-core with previous version of Cairo.

Another issue appeared when building with previous versoin of Cairo,
that the dumb version of functions triggered two warnings. No type
signature and unused variables. Also fixed.